Start development 30-08-2023

+++ database/migrations/2023_02_21_181827_create_movements_table.php
@@ -17,7 +17,7 @@ return new class extends Migration {
             $table->enum('recurrence', ['Monthly', 'Weekly', 'Daily']);
             $table->dateTime('created_at')->index();
             $table->dateTime('updated_at')->index();
-            $table->unsignedInteger('created_by')->index();
+            $table->unsignedBigInteger('created_by')->index();
 
             // foreign keys
             $table->foreign('created_by')->references('id')->on('users');
